alert for some things that they arent supported by in-memory yet

+++ chroma-server/chroma_server/index/hnswlib.py
@@ -29,15 +29,16 @@ class Hnswlib(Index):
     def set_save_folder(self, save_folder):
         self._save_folder = save_folder
 
+    def get_save_folder(self):
+        return self._save_folder
+
     def run(self, model_space, uuids, embeddings, space='l2', ef=10, num_threads=4):
         # more comments available at the source: https://github.com/nmslib/hnswlib
         dimensionality = len(embeddings[0])
         
-        
         for uuid, i in zip(uuids, range(len(uuids))):
             self._id_to_uuid[i] = str(uuid)
             self._uuid_to_id[str(uuid)] = i
-
 
         index = hnswlib.Index(space=space, dim=dimensionality) # possible options are l2, cosine or ip
         index.init_index(max_elements=len(embeddings), ef_construction=100, M=16) 
@@ -120,5 +121,6 @@ class Hnswlib(Index):
         return uuids, distances
 
     def reset(self):
-        for f in os.listdir(f'{self._save_folder}'):
-            os.remove(os.path.join(f'{self._save_folder}', f))
+        if os.path.exists(f'{self._save_folder}'):
+            for f in os.listdir(f'{self._save_folder}'):
+                os.remove(os.path.join(f'{self._save_folder}', f))
